Number_Extractor-main/requirements.txt
langchain==0.2.2
google
langchain-google-genai
langchain_community
python-dotenv
streamlit

#pip install spacy
#python -m spacy download en_core_web_sm
